# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mithqal — Constitutional Settlement Institution API",
    "version": "v19.0-stable",
    "description": "The Mithqal public API. All endpoints are relative to the site root (https://mithqal.vercel.app). The Admin endpoints require an authenticated operator session (NextAuth JWT). Rate limiting: 5 submissions/hour/IP on public POST endpoints.",
    "contact": {
      "name": "Mithqal Formation Committee",
      "url": "https://mithqal.vercel.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/MITHQALMTQ/mithqal/blob/main/LICENSE"
    }
  },
  "servers": [
    { "url": "https://mithqal.vercel.app", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/admin/smtp-test": {
      "get": {
        "tags": ["Admin"],
        "summary": "SMTP configuration status",
        "description": "Returns the current SMTP configuration status (configured/not configured) WITHOUT exposing credentials.",
        "security": [{ "nextauth": [] }],
        "responses": {
          "200": {
            "description": "SMTP status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "configured": { "type": "boolean" },
                    "host": { "type": "string" },
                    "port": { "type": "string" },
                    "user": { "type": "string" },
                    "notifyEmail": { "type": "string" },
                    "passConfigured": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Send SMTP test email",
        "description": "Sends a test email to ADMIN_NOTIFY_EMAIL via the configured SMTP transporter. Verifies end-to-end connectivity (auth + STARTTLS + delivery).",
        "security": [{ "nextauth": [] }],
        "responses": {
          "200": { "description": "Email sent" },
          "500": { "description": "SMTP not configured or send failed" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "nextauth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth JWT session cookie. Obtain via POST /api/auth/callback/credentials."
      }
    }
  }
}